+++ condow_core/src/lib.rs
@@ -69,7 +69,7 @@
 //! use futures::AsyncSeekExt;
 //! let mut reader = condow.blob().at("a location").trusted_blob_size(13).random_access_reader().finish().await.unwrap();
 //! let mut buf = Vec::new();
-//! reader.seek(std::io::SeekFrom::Start(3));
+//! reader.seek(std::io::SeekFrom::Start(2));
 //! reader.read_to_end(&mut buf).await.unwrap();
 //! assert_eq!(buf, b"remote BLOB");
 //! # }