+++ condow_http/examples/simple.rs
@@ -9,10 +9,7 @@ use condow_http::HttpClient;
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> Result<(), AnyError> {
-    let config = Config::default()
-        .buffer_size(1)
-        .part_size_bytes(1000)
-        .always_get_size(true);
+    let config = Config::default().buffer_size(1).part_size_bytes(1000);
     let condow = HttpClient::condow(config)?;
 
     let url = "https://gist.githubusercontent.com/nyurik/88730133a8d00ead67ac8520640e1fc1/raw/29459acf2aef3281489d0eb1b09f491831cf2324/denormalize_osm_data.md";